item admin - securing the views

+++ services/views.py
@@ -1,6 +1,6 @@
 from django.contrib import messages
 from django.shortcuts import (render, get_object_or_404, redirect, reverse)
-
+from django.contrib.auth.decorators import login_required
 
 from .models import Item, Category
 from .forms import ItemForm
@@ -44,8 +44,14 @@ def single_item(request, item_id):
     return render(request, 'services/single_service.html', context)
 
 
+
+@login_required
 def add_item(request):
     """ Add a item to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry only website owner can do that.')
+        return redirect(reverse('home'))
+    
     if request.method == 'POST':
         form = ItemForm(request.POST, request.FILES)
         if form.is_valid():
@@ -65,8 +71,13 @@ def add_item(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_item(request, item_id):
     """ Edit item in the store """
+    if not request.use.is_superuser:
+        messages.error(request, 'Sorry only website owner can do that.')
+        return redirect(reverse('home'))
+    
     item = get_object_or_404(Item, pk=item_id)
     if request.method == 'POST':
         form = ItemForm(request.POST, request.FILES, instance=item)
@@ -89,8 +100,13 @@ def edit_item(request, item_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_item(request, item_id):
     """ Delete item from the store """
+    if not request.use.is_superuser:
+        messages.error(request, 'Sorry only website owner can do that.')
+        return redirect(reverse('home'))
+    
     item = get_object_or_404(Item, pk=item_id)
     item.delete()
     messages.success(request, 'Item deleted!')
